Lock the thread during to commit actions to mutations in store
